+++ test/Routines/PRM/IS-PRM-SURVEY/buildPrecursorTable.jl
@@ -62,11 +62,11 @@
     0x00000001 │ Peptide("PEPTIDE", 0x00000001, Set{UInt32}())
     """
 
-    test_mods::Dict{String, Float32} = 
-    Dict{String, Float32}(
-        "Carb" => Float32(57.021464),
-        "Harg" => Float32(10),
-        "Hlys" => Float32(8),
+    test_mods::Dict{String, Float64} = 
+    Dict{String, Float64}(
+        "Carb" => Float64(57.021464),
+        "Harg" => Float64(10),
+        "Hlys" => Float64(8),
     )
     fixed_mods = [(p=r"C", r="C[Carb]")]
     var_mods = [(p=r"(K$)", r="[Hlys]"), (p=r"(R$)", r="[Harg]")]
